Fetch items when there's no cache for the Entity.
Change KVCFetching behaviour when cache is used, but no instances cache is found for the wanted entity.

The previous behaviour was to return nil, as for a cache miss when the instances cache is available for the entity.
We now fallback to a regular fetch request.

This fixes issues where the cache was created according the classes of objects being imported,
and some external objects were referenced in relationship. (i.e., not sideloaded.)
The previous behaviour made KVCMapping to break the relationship, even when we actually had both objects locally.